Refactors and deploys Docusaurus website
Migrates from the old documentation website to a new Docusaurus implementation, unifying documentation and website under the same domain.

This change updates the GitHub Pages deployment workflow to use the `main` branch and Node.js v18.  It also configures Docusaurus with a new theme, custom CSS, and a homepage to provide a better user experience.